+++ previews/elements/delete_component_preview.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+class Elements::DeleteComponentPreview < ViewComponent::Preview
+  def default
+    render Bulma::DeleteComponent.new
+  end
+
+  def sizes
+  end
+end